requirements.txt
asgiref==3.7.2
boto3==1.28.84
botocore==1.31.84
Django==3.2.23
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
